+++ src/compiler/llvm_codegen_builtins.c
@@ -714,11 +714,9 @@ void llvm_emit_builtin_call(GenContext *c, BEValue *result_value, Expr *expr)
 			llvm_emit_compare_exchange(c, result_value, expr);
 			return;
 		case BUILTIN_FRAMEADDRESS:
+		case BUILTIN_RETURNADDRESS:
 		{
-			LLVMTypeRef type[2] = { c->ptr_type, llvm_get_type(c, type_int) };
-			LLVMValueRef value = LLVMConstNull(type[1]);
-			value = llvm_emit_call_intrinsic(c, intrinsic_id.frameaddress, type, 1, &value, 1);
-			llvm_value_set(result_value, value, expr->type);
+			llvm_emit_simple_builtin(c, result_value, expr, intrinsic_id.frameaddress);
 			return;
 		}
 		case BUILTIN_SELECT: