+++ src/compiler/llvm_codegen_internal.h
@@ -204,6 +204,7 @@ typedef struct
 	unsigned powi;
 	unsigned prefetch;
 	unsigned readcyclecounter;
+	unsigned returnaddress;
 	unsigned rint;
 	unsigned round;
 	unsigned roundeven;